pyproject.toml
[project]
name = "website-summarization-app"
version = "0.1.0"
description = "Website summarization app with Resonate, Flask, and Ollama."
authors = [
    { name = "Cully Wakelin", email = "flossypurse@gmail.com" }
]
dependencies = [
    "resonate-sdk>=0.4.12",
    "flask>=3.0.3",
    "charset-normalizer>=3.4.0",
    "selenium>=4.26.1",
    "bs4>=0.0.2",
    "ollama>=0.3.3",
    "types-beautifulsoup4>=4.12.0.20241020",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
    "gateway" = "src.gateway:main"
    "app" = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.13.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
